+++ src/imio/history/interfaces.py
@@ -12,9 +12,10 @@ class IImioHistoryLayer(IBrowserRequest):
 class IImioHistory(Interface):
     """ """
 
-    def getHistory(self):
-        """Returns the history for given p_obj, sorted in reverse order
-        (most recent change first)."""
+    def getHistory(self, checkMayView=True):
+        """Returns the history for context, sorted in reverse order
+        (most recent change first).  If p_checkMayView is True (default),
+        the method 'mayViewComment' is called on every history event."""
 
     def historyLastEventHasComments(self):
         """Returns True if the last event of the object's history has a comment."""