+++ src/imio/history/tests/test_contenthistory.py
@@ -103,3 +103,11 @@ class TestContentHistory(IntegrationTestCase):
         self.assertTrue(lastEvent['action'] == 'publish')
         # now comment is no more viewable
         self.assertTrue(lastEvent['comments'] == HISTORY_COMMENT_NOT_VIEWABLE)
+
+        # getHistory can be called with checkMayView set to False,
+        # in this case, mayViewComment check is not done
+        history = view.getHistory(checkMayView=False)
+        lastEvent = history[0]
+        self.assertTrue(lastEvent['action'] == 'publish')
+        # comment is viewable as mayViewComment was not done
+        self.assertTrue(lastEvent['comments'] == 'My comment')